Move idetify actions in mixin
