+++ addon/templates/components/flexberry-identify-panel.hbs
