+++ tests/dummy/app/templates/map.hbs
@@ -175,7 +175,7 @@
           activate=(action "onMapToolActivate" target=mapToolbar)
           bufferActive=bufferActive
           bufferUnits=bufferUnits
-          bufferRadius=bufferRadius          
+          bufferRadius=bufferRadius
           layerMode=identifyLayersOption
           toolMode=identifyToolOption
           leafletMap=leafletMap
